Change project structure to an area based format
This commit is focused on restructuring the project to an area based approach,
which is partially inspired by the tree structure of html, to define how the
screen will be layed out. termgraph.go has been removed in favor of control.go,
and the functionality of the last version has mostly been added for this new
approach. Areas will allow the user to work in local coordinate spaces, print
strings cleanly, and overall make more complicated user interfaces that would
not normally be possible with the last approach.

The next steps are to design ang implement updating functionality so that the
screen can actually be updated.